UNT0039: do not report GetComponent inside editor-only messages
GetComponent calls in Reset/OnValidate are typically used to populate
serialized fields with default references that a user can override, so
the component is not a hard requirement (#453).

Reuse the editor-only message detection already used by UNT0040 by
moving it to SyntaxNodeExtensions.
