+++ doc/UNT0039.md
@@ -38,3 +38,21 @@ public class PlayerScript : MonoBehaviour
 ```
 
 A code fix is offered for this diagnostic to automatically apply this change.
+
+## Examples of patterns that are not flagged by this analyzer
+
+`GetComponent` calls inside editor-only messages (`Reset`, `OnValidate`) are not flagged: those messages are typically used to populate serialized fields with default references that a user can override, so the component is not a hard requirement.
+
+```csharp
+using UnityEngine;
+
+public class PlayerScript : MonoBehaviour
+{
+    public Rigidbody rb;
+
+    void Reset()
+    {
+        rb = GetComponent<Rigidbody>();
+    }
+}
+```
